refactor: archive the util & service layer
the service abstraction layer i had implemented a couple commits ago has
proved to be an unnecessary overengineering for the current scope of the
project and therefore is being removed and archived. back to the good
old days of regular prisma queries bloating up the controllers! that is
of course for the better. ¯\_(ツ)_/¯