feat: add file upload to api/admin/upload endpoint
the admin will be able to upload one image per post (due to storage
limitations on free supabase plan). this might be improved upon in the
future.

the upload procedure uses multer to multipart/form-data parsing and the
storage client from supabase-js to upload the file to a supabase storage