feat: implement post creation logic
the posts are now created and being sent to the database. however, there
are a lot of improvements to be made. currently, it only accepts the
following fields:

- title,
- content,
- excerpt,

and it then generates the id, timestamps etc. the next commit will aim
to implement the image uploading, supabase.storage processes to this.