chore: remove unused packages & rename frontend
the packages `packages/frontend-admin` & `packages/shared` have been
removed altogether as the idea to have a separate frontend for the admin
panel was scraped along the way. this change also necessitated renaming
the `packages/frontend-blog` to simply `packages/frontend` hoping this
change wont break anything important. did my best to prevent that, but
in any case, i will be backing up a copy of the main branch before
merging just to be on the safe side.