feat: add `StartupGate` & wake up backend
the render PaaS platform spins down the server on free tier accounts as
a means to incentivize users to upgrade to a paid tier. this causes an
arbitrary amount (50 seconds or more as per the render docs) of waiting
before the server is 'warmed up'. during that time, the user is able to
view the frontend (deployed on netlify) but unable to interact with the
backend features, which gives the impression of the app being broke. to
prevent that, i saw fit to implement a bridge, sort of a gate to init
the server on render platform and in the mean time, display information
on the frontend regarding this; as well as a fallback screen in case
something goes wrong.