feat: change email provider to `Resend`
apparently, the render PaaS blocks the SMTP mailing, causeing the email
to never be sent on contact form. this commit changes the provider to an
actual transactional email provider with an API key so that this issue
is worked around. however, since this email service provider (along
with many others) do actually require a verified domain, this branch is
to be postponed until that constraint is met.