feat: implement a working command palette
the command palette now works in searching through posts, categories and
tags, furthermore, currently, it supports to navigate to the profile of
the current user. further functionalities will be implemented to the
command palette, such as navigating throughout the site, settings etc.

thid current implementation is but a proof of concept.